booltypes {bit}	R Documentation

Boolean types

Description

The ordered factor booltypes ranks the boolean types.

Usage

booltypes

Format

An object of class ordered (inherits from factor) of length 6.

Details

There are currently six boolean types, booltypes is an ordered vector with the following ordinal levels

nobool

non-boolean types

logical

for representing any boolean data including NA

bit

for representing dense boolean data

bitwhich

for representing sparse (skewed) boolean data

which

for representing sparse boolean data with few TRUE

ri

range-indexing, for representing sparse boolean data with a single range of TRUE

booltypes has a names attribute such that elements can be selected by name.

Note

do not rely on the internal integer codes of these levels, we might add-in hi later

See Also

booltype, is.booltype, as.booltype

[Package bit version 4.0.5 Index]