in.bitwhich {bit}	R Documentation

Check existence of integers in table

Description

If the table is sorted, this can be much faster than %in%

Usage

in.bitwhich(x, table, is.unsorted = NULL)

Arguments

x
a vector of integer

table
a bitwhich object or a vector of integer

is.unsorted
logical telling the function whether the table is (un)sorted. With the defautl NULL FALSE is assumed for bitwhich tables, otherwise TRUE

Value

logical vector

See Also

%in%

Examples

x <- bitwhich(100)
x[3] <- TRUE
in.bitwhich(c(NA,2,3), x)

[Package bit version 4.0.4 Index]