maxindex.default {bit}	R Documentation

Get maxindex (length of boolean vector) and poslength (number of 'selected' elements)

Description

For is.booltype objects the term length is ambiguous. For example the length of which corresponds to the sum of logical. The generic maxindex gives length(logical) for all booltypes. The generic poslength gives the number of positively selected elements, i.e. sum(logical) for all booltypes (and gives NA if NAs are present).

Usage

## Default S3 method:
maxindex(x, ...)

## Default S3 method:
poslength(x, ...)

## S3 method for class 'logical'
maxindex(x, ...)

## S3 method for class 'logical'
poslength(x, ...)

## S3 method for class 'bit'
maxindex(x, ...)

## S3 method for class 'bit'
poslength(x, ...)

## S3 method for class 'bitwhich'
maxindex(x, ...)

## S3 method for class 'bitwhich'
poslength(x, ...)

## S3 method for class 'which'
maxindex(x, ...)

## S3 method for class 'which'
poslength(x, ...)

## S3 method for class 'ri'
maxindex(x, ...)

## S3 method for class 'ri'
poslength(x, ...)

maxindex(x, ...)

poslength(x, ...)

Arguments

x
an R object, typically a is.booltype object.

...
further arguments (ignored)

Value

an integer scalar

Methods (by class)

default: default method for maxindex

default: default method for poslength

logical: maxindex method for class logical

logical: poslength method for class logical

bit: maxindex method for class bit

bit: poslength method for class bit

bitwhich: maxindex method for class bitwhich

bitwhich: poslength method for class bitwhich

which: maxindex method for class which

which: poslength method for class which

ri: maxindex method for class ri

ri: poslength method for class ri

Examples

r <- ri(1,2,12)
i <- as.which(r)
w <- as.bitwhich(r)
b <- as.bit(r)
l <- as.logical(r)
u <- which(l)      # unclassed which

sapply(list(r=r,u=u,i=i,w=w,b=b,l=l), function(x){
  c(length=length(x), sum=sum(x), maxindex=maxindex(x), poslength=poslength(x))
})

[Package bit version 4.0.5 Index]