range_nanozero {bit}	R Documentation

Remove zeros and get range and number of NAs

Description

Remove zeros and get range and number of NAs

Usage

range_nanozero(x)

Arguments

x
an integer vector

Value

an integer vector without zeros and with an attribute range_na with three elements

1

min integer

2

max integer

3

number of NAs

See Also

range_na and range_sortna

Examples

range_nanozero(c(0L,1L,2L,NA))

[Package bit version 4.0.4 Index]