as.bit.NULL {bit}	R Documentation

Coercing to bit

Description

Coercing to bit vector

Usage

## S3 method for class ''NULL''
as.bit(x, ...)

## S3 method for class 'bit'
as.bit(x, ...)

## S3 method for class 'logical'
as.bit(x, ...)

## S3 method for class 'integer'
as.bit(x, ...)

## S3 method for class 'double'
as.bit(x, ...)

## S3 method for class 'bitwhich'
as.bit(x, ...)

## S3 method for class 'which'
as.bit(x, length = attr(x, "maxindex"), ...)

## S3 method for class 'ri'
as.bit(x, ...)

as.bit(x = NULL, ...)

Arguments

x
an object of class bit, logical, integer, bitwhich or an integer from as.which or a boolean ff

...
further arguments

length
the length of the new bit vector

Details

Coercing to bit is quite fast because we use a double loop that fixes each word in a processor register

Value

is.bit returns FALSE or TRUE, as.bit returns a vector of class 'bit'

Methods (by class)

NULL: method to coerce to bit (zero length) from NULL

bit: method to coerce to bit from bit

logical: method to coerce to bit from logical

integer: method to coerce to bit from integer (0L and NA become FALSE, everthing else becomes TRUE)

double: method to coerce to bit from double (0 and NA become FALSE, everthing else becomes TRUE)

bitwhich: method to coerce to bit from bitwhich

which: method to coerce to bit from which

ri: method to coerce to bit from ri

Note

Zero is coerced to FALSE, all other numbers including NA are coerced to TRUE. This differs from the NA-to-FALSE coercion in package ff and may change in the future.

Author(s)

Jens Oehlschlägel

See Also

CoercionToStandard, as.booltype, as.bit, as.bitwhich , as.which, as.ri, as.hi, as.ff

Examples

as.bit(c(0L,1L,2L,-2L,NA))
as.bit(c(0,1,2,-2,NA))

as.bit(c(FALSE, NA, TRUE))

[Package bit version 4.0.4 Index]