bit-package {bit}	R Documentation

bit: Classes and methods for fast memory-efficient boolean selections

Description

Provided are classes for boolean and skewed boolean vectors, fast boolean methods, fast unique and non-unique integer sorting, fast set operations on sorted and unsorted sets of integers, and foundations for ff (range indices, compression, chunked processing).

Details

For details view the vignettes ../doc/bit-usage.pdf and ../doc/bit-performance.pdf

[Package bit version 4.0.4 Index]