chunk {bit}	R Documentation

Methods for chunked range index

Description

Calls chunks to create a sequence of range indexes along the object which causes the method dispatch.

Usage

chunk(x = NULL, ...)

## Default S3 method:
chunk(x = NULL, ..., RECORDBYTES = NULL, BATCHBYTES = NULL)

Arguments

x
the object along we want chunks

...
further arguments passed to chunks

RECORDBYTES
integer scalar representing the bytes needed to process a single element of the boolean vector (default 4 bytes for logical)

BATCHBYTES
integer scalar limiting the number of bytes to be processed in one chunk, default from getOption("ffbatchbytes") if not null, otherwise 16777216

Details

chunk is generic, the default method is described here, other methods that automatically consider RAM needs are provided with package 'ff', see for example chunk.ffdf

Value

returns a named list of ri objects representing chunks of subscripts

Methods (by class)

default: default vector method

available methods

chunk.default, chunk.ff_vector, chunk.ffdf

Author(s)

Jens Oehlschlägel

See Also

chunks, ri, seq, bbatch

Examples

  chunk(complex(1e7))
  chunk(raw(1e7))
  chunk(raw(1e7), length=3)
  
  chunks(1,10,3)
  # no longer do
  chunk(1,100,10)
  # but for bckward compatibility this works
  chunk(from=1,to=100,by=10)
  

[Package bit version 4.0.4 Index]