ri {bit}	R Documentation

Range index

Description

A range index can be used to extract or replace a continuous ascending part of the data

Usage

ri(from, to = NULL, maxindex = NA)

## S3 method for class 'ri'
print(x, ...)

Arguments

from
first position

to
last posistion

maxindex
the maximal length of the object-to-be-subscripted (if known)

x
an object of class 'ri'

...
further arguments

Value

A two element integer vector with class 'ri'

Author(s)

Jens Oehlschlägel

See Also

as.hi

Examples

 bit(12)[ri(1,6)]

[Package bit version 4.0.4 Index]