rlepack {bit}	R Documentation

Hybrid Index, rle-pack utilities

Description

Basic utilities for rle packing and unpacking and apropriate methods for rev and unique.

Usage

rlepack(x, ...)

## S3 method for class 'integer'
rlepack(x, pack = TRUE, ...)

rleunpack(x)

## S3 method for class 'rlepack'
rleunpack(x)

## S3 method for class 'rlepack'
rev(x)

## S3 method for class 'rlepack'
unique(x, incomparables = FALSE, ...)

## S3 method for class 'rlepack'
anyDuplicated(x, incomparables = FALSE, ...)

Arguments

x
in 'rlepack' an integer vector, in the other functions an object of class 'rlepack'

...
just to keep R CMD CHECK quiet (not used)

pack
FALSE to suppress packing

incomparables
just to keep R CMD CHECK quiet (not used)

Value

A list with components

first
the first element of the packed sequence

dat
either an object of class rle or the complete input vector x if rle-packing is not efficient

last
the last element of the packed sequence

Author(s)

Jens Oehlschlägel

See Also

hi, intrle, rle, rev, unique

Examples

  x <- rlepack(rep(0L, 10))

[Package bit version 4.0.4 Index]